Saas multidomain v01

+++ init.sh
@@ -2,6 +2,12 @@
 
 # disable paritcular supervisor job by deleting jobs files 
 
+if [ -n "$MULTI_DOMAINS" ]
+then
+  ./init_multidomains.sh
+  exit 0;
+fi
+
 if [ "$DISBALE_PHP_FPM" == 'true' ]
 then
     rm -f /etc/supervisor/conf.d/php-fpm.conf
@@ -27,6 +33,10 @@ then
 fi
 
 # set env from `LARAVEL_` prefixed env vars
+# this also setup MULTI_DOMAINS eg 
+# when MULTI_DOMAINS: "api-sprawnymarketing.escolalms.com,api-gest.escolalms.com" 
+# then API_SPRAWNYMARKETING_ESCOLALMS_COM_APP_NAME: '"Sprawny Marketing"'
+
 php docker/envs/envs.php
 
 # if binded by k8s or docker those folders might need to be recreated
@@ -39,7 +49,8 @@ mkdir storage/app
 mkdir storage/logs
 
 # run all laravel related tasks 
-
+# klucze mozna trzymac jako zmienne srodowiskowe wiec .... 
+# https://github.com/gecche/laravel-multidomain/issues/51
 # create keys from env base64 variables 
 if [ -n "$JWT_PUBLIC_KEY_BASE64" ]; then
     echo "Storing public RSA key for JWT generation - storage/oauth-public.key"
@@ -84,72 +95,4 @@ fi
 php artisan storage:link --force --no-interaction
 php artisan h5p:storage-link
 
-# MULTI_DOMAINS
-if [ -n "$MULTI_DOMAINS" ]; then
-  IFS=',' read -ra domains <<< "$MULTI_DOMAINS"
 
-  for domain in "${domains[@]}"; do
-    echo "$domain"
-
-    # horizon
-    if [ -z "$DISBALE_HORIZON" ] || [ "$DISBALE_HORIZON" != "true" ];
-    then
-      cp "docker/conf/supervisor/example/horizon.conf.example" "/etc/supervisor/custom.d/horizon.$domain.conf"
-      sed "s/\$HORIZON_DOMAIN/$domain/g" "docker/conf/supervisor/example/horizon.conf.example" > "/etc/supervisor/custom.d/horizon.$domain.conf"
-    else
-      echo "Horizon disabled"
-    fi
-    # scheduler
-    if [ -z "$DISBALE_SCHEDULER" ] || [ "$DISBALE_SCHEDULER" != "true" ];
-    then
-      cp "docker/conf/supervisor/example/scheduler.conf.example" "/etc/supervisor/custom.d/scheduler.$domain.conf"
-      sed "s/\$SCHEDULER_DOMAIN/$domain/g" "docker/conf/supervisor/example/scheduler.conf.example" > "/etc/supervisor/custom.d/scheduler.$domain.conf"
-    else
-      echo "Schedule disabled"
-    fi
-
-    DOMAIN_KEY=$(echo "$domain" | tr '[:lower:]' '[:upper:]')
-    DOMAIN_KEY=$(echo "$DOMAIN_KEY" | tr '.-' '__')
-
-    DB_NAME_KEY="${DOMAIN_KEY}_DB_DATABASE"
-    DB_NAME_VALUE=${!DB_NAME_KEY}
-
-    # create db if not exists
-    if PGPASSWORD="$POSTGRES_PASSWORD" psql -h "$LARAVEL_DB_HOST" -U "$POSTGRES_USER" -d "$DB_NAME_VALUE" -c ";" >/dev/null 2>&1; then
-      echo "DB $DB_NAME_VALUE already exists"
-    else
-      PGPASSWORD="$POSTGRES_PASSWORD" psql -h "$LARAVEL_DB_HOST" -U "$POSTGRES_USER" -c "CREATE DATABASE $DB_NAME_VALUE;"
-      echo "DB $DB_NAME_VALUE has been created"
-    fi
-
-    # db migrate
-    if [ "$DISBALE_DB_MIGRATE" == 'true' ]
-    then
-        echo "Disable db migrate"
-    else
-        php artisan migrate --force --domain=$domain
-    fi
-
-    # db seed
-    if [ "$DISBALE_DB_SEED" == 'true' ]
-    then
-        echo "Disable db:seed"
-    else
-        php artisan db:seed --domain=$domain --class=PermissionsSeeder --force --no-interaction
-    fi
-
-    # storage
-    STORAGE_DIRECTORY=$(echo "$domain" | tr '[:upper:]' '[:lower:]' | tr '.' '_')
-    STORAGE_PUBLIC_KEY="${DOMAIN_KEY}_APP_PUBLIC_STORAGE"
-    STORAGE_PUBLIC_NAME=${!STORAGE_PUBLIC_KEY}
-    if [ -n "$STORAGE_PUBLIC_NAME" ]; then
-      ln -s /var/www/html/storage/${STORAGE_DIRECTORY}/app/public public/storage${STORAGE_PUBLIC_NAME}
-      ln -s /var/www/html/storage/${STORAGE_DIRECTORY}/app/h5p public/h5p${STORAGE_PUBLIC_NAME}
-    fi
-
-    php artisan optimize:clear --domain=$domain
-
-  done
-else
-  echo "Environment variable MULTI_DOMAINS is empty."
-fi
